Refactor tests into unit tests sub-package
